move schema files to the package with other schemas
